:recycle: Extract context managers
