+++ kivy_ios/recipes/itsdangerous/__init__.py
@@ -1,5 +1,6 @@
 # pure-python package, this can be removed when we'll support any python package
-from kivy_ios.toolchain import PythonRecipe, shprint, cd
+from kivy_ios.toolchain import PythonRecipe, shprint
+from kivy_ios.context_managers import cd
 from os.path import join
 import sh
